pyproject.toml
[tool.poetry]
name = "strongunwrittenbetaversion"
version = "0.1.0"
description = ""
authors = ["RosaleeKnight <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
